pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-cinema"
dynamic = ["version"]
description = "Audio processing control system with CamillaDSP integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Erwan Daniel"},
]
keywords = ["audio", "dsp", "camilladsp", "pulseaudio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Django>=6.0",
    "pulsectl>=24.12.0",
    "PyYAML>=6.0",
    "websocket-client>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/open-cinema"
Documentation = "https://github.com/yourusername/open-cinema#readme"
Repository = "https://github.com/yourusername/open-cinema"
Issues = "https://github.com/yourusername/open-cinema/issues"

[tool.setuptools.dynamic]
version = {attr = "version.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["opencinema*", "api*", "core*", "plugin*", "version"]
exclude = ["tests*", ".venv*", "deployment*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.venv2
  | build
  | dist
  | migrations
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "opencinema.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["tests"]
